[CodeCompletion] Move CompletionOverrideLookup to its own file
